refactor: Clean-up app folder structure
