--- frontend/src/app/app.component.scss
@@ -1,15 +1,0 @@
-.container {
-    max-width: 1170px;
-    margin: auto;
-    position: relative;
-    padding-left: 15px;
-    padding-right: 15px;
-    @media(min-width: 400px) {
-        padding-left: 20px;
-        padding-right: 20px;
-    }
-}
-
-.mat-progress-bar {
-    z-index: 20;
-}
