+++ frontend/src/app/misc/SentryErrorHandler.ts
